refactor: split utils.js, remove UnoCSS, add structured logging
Architecture:
- Split utils.js God Object into 7 focused lib modules (database, keys,
  logger, prefixes, rate-limit, responses, user, validation)
- Fix N+1 query in listUsers() with batch URL signing

Code Quality:
- Convert 5 broken test files from CommonJS to ESM
- Add structured JSON logging with LOG_LEVEL support
- Add JSDoc type annotations with @ts-check to lib modules
- Standardize service filenames to kebab-case

Problem Fit:
- Remove UnoCSS (redundant with Tailwind+DaisyUI)
- Replace 18 UnoCSS icon classes with svelte-hero-icons components
- Remove 5 unused dependencies (@iconify-json/*, @unocss/*, unocss)
