+++ backend/lambdas/api/lib/prefixes.js
@@ -1,0 +1,20 @@
+/**
+ * Key prefixes for single-table DynamoDB design
+ * @module lib/prefixes
+ */
+
+const PREFIX = {
+  USER: 'USER#',
+  COMMENT: 'COMMENT#',
+  CONV: 'CONV#',
+  MSG: 'MSG#',
+  REACTION: 'REACTION#',
+  RATE: 'RATE#',
+  LETTER: 'LETTER#',
+  VERSION: 'VERSION#',
+  DRAFT: 'DRAFT#',
+}
+
+module.exports = {
+  PREFIX,
+}
