+++ package.json
@@ -26,15 +26,17 @@
     "aws-sdk": "^2.1692.0",
     "gray-matter": "^4.0.3",
     "jose": "^6.0.12",
-    "marked": "^15.0.7"
+    "marked": "^15.0.7",
+    "svelte-hero-icons": "^5.2.0"
+  },
+  "optionalDependencies": {
+    "sharp": "^0.31.3"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^3.5.1",
     "@aws-sdk/client-dynamodb": "^3.940.0",
     "@aws-sdk/lib-dynamodb": "^3.940.0",
     "@aws-sdk/s3-request-presigner": "^3.940.0",
-    "@iconify-json/heroicons-outline": "^1.1.10",
-    "@iconify-json/heroicons-solid": "^1.1.11",
     "@importantimport/eslint-config": "1.0.0-alpha.2",
     "@playwright/test": "^1.56.1",
     "@sveltejs/adapter-auto": "^3.1.1",
@@ -48,8 +50,6 @@
     "@types/node": "^20.11.17",
     "@types/sanitize-html": "^2.16.0",
     "@types/unist": "^3.0.2",
-    "@unocss/eslint-plugin": "^0.62.3",
-    "@unocss/extractor-svelte": "^0.55.7",
     "@vite-pwa/sveltekit": "^1.1.0",
     "@vitest/coverage-v8": "^2.1.0",
     "artillery": "^2.0.26",
@@ -83,7 +83,6 @@
     "tslib": "^2.6.2",
     "typescript": "^5.3.3",
     "unist-util-visit": "^4.1.2",
-    "unocss": "^0.58.5",
     "uuid": "^13.0.0",
     "vite": "^5.1.1",
     "vite-imagetools": "^4.0.19",
