+++ vite.config.ts
@@ -3,12 +3,10 @@ import { SvelteKitPWA as pwa } from '@vite-pwa/sveltekit'
 // @ts-expect-error ts(7016)
 import LightningCSS from 'postcss-lightningcss'
 import TailwindCSS from 'tailwindcss'
-import unoCSS from 'unocss/vite'
 import { defineConfig } from 'vite'
 import { imagetools } from 'vite-imagetools'
 
 import tailwindConfig from './tailwind.config'
-import unoConfig from './uno.config'
 
 export default defineConfig({
   css: {
@@ -18,7 +16,6 @@ export default defineConfig({
   },
   envPrefix: 'URARA_',
   plugins: [
-    unoCSS(unoConfig),
     imagetools(),
     sveltekit(),
     pwa({
