+++ vitest.config.ts
@@ -15,12 +15,6 @@ export default defineConfig({
       '.svelte-kit',
       'tests/integration/**',
       '**/.aws-sam/**',
-      // TODO: These tests need updating for the consolidated API structure
-      'tests/unit/comments-handler.test.js',
-      'tests/unit/messages-handler.test.js',
-      'tests/unit/profile-handler.test.js',
-      'tests/unit/profile-security.test.js',
-      'tests/unit/reactions-handler.test.js',
     ],
     coverage: {
       provider: 'v8',
